Fix associations for rails 5
This is a test-project, so not all associations
were completely specified.

Since rails 5 belongs_to is by default required. So
for nested attributes to be able to save, we need to
specify for which inverse association (since the id
is not yet set at validation time).

For consistency, I always added the `belongs_to`
relation were missing (and thus also the inverse_of).